requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
phonenumbers==8.13.19
twilio==8.5.0
pandas==2.0.3
gunicorn==21.2.0
Werkzeug==2.3.7
WTForms==3.0.1